Rename the header to fix a typo
